third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/editing/inserting/delete-insignificant-text-crash-expected.txt

CONSOLE WARNING: Listener added for a 'DOMCharacterDataModified' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
This tests deleting a node in DOMCharacterDataModified doesn't result in a crash.

PASS
